fastapi_microservice/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
qdrant-client>=1.7.0
google-generativeai>=0.3.0

